Moved `SpriteSheetAnimationFrame` to `MonoGame.Extended.Graphics`
